// scripts/Pix-Scripting/src/types.ts
export interface PixivImageUrls {
  square_medium?: string
  medium?: string
  large?: string
  original?: string
}

export interface PixivTag {
  name: string
  translated_name?: string | null
}

export interface PixivUser {
  id: number
  name: string
  account: string
  profile_image_urls?: { medium?: string }
  comment?: string
  is_followed?: boolean
  is_mypixiv?: boolean
}

export interface PixivSeries {
  id: number
  title?: string | null
}
// ...
export interface PixivNovel {
  id: number
  title: string
  caption?: string
  user: PixivUser
  tags: PixivTag[]
  create_date: string
  page_count: number
  x_restrict: number
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  is_muted: boolean
  novel_ai_type?: number
  total_comments: number
  text_length?: number
  visible: boolean
  series?: PixivSeries | null
  episode_number?: number
  novel_marker?: {
    page: number
  } | null
  image_urls?: {
    square_medium?: string
    medium?: string
    large?: string
  } | null
  cover?: {
    urls?: { "240mw"?: string; "480mw"?: string; "1200x1200"?: string }
  } | null
}


export interface PixivNovelMarker {
  id: number
  novel: PixivNovel
  novel_marker: {
    page: number
  }
}

export interface TextEmbeddedImage {
  novelImageId?: string
  sl?: string
  urls: {
    "128x128"?: string
    "240mw"?: string
    "480mw"?: string
    "1200x1200"?: string
    original?: string
  }
}

export interface PixivNovelDetail {
  id: number
  title: string
  user: PixivUser
  tags: PixivTag[]
  caption?: string
  text_length?: number
  create_date: string
  page_count: number
  x_restrict: number
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  total_comments: number
  cover?: { urls?: { "240mw"?: string; "480mw"?: string; "1200x1200"?: string } }
  image_urls?: PixivImageUrls | null
  textEmbeddedImages?: Record<string, TextEmbeddedImage>
  content: string
  episode_number?: number
  novel_marker?: {
    page: number
  } | null
  series?: {
    id: number
    title?: string
  } | null
  series_next?: {
    id: number
    title?: string
    url?: string
  } | null
  series_prev?: {
    id: number
    title?: string
    url?: string
  } | null
}